Added pin for chat start
