+++ logicle/app/chat/components/StartChatFromHere.tsx
@@ -2,6 +2,7 @@ import { useTranslation } from 'next-i18next'
 import { IconSend } from '@tabler/icons-react'
 import { Avatar } from '@/components/ui/avatar'
 import * as dto from '@/types/dto'
+import { AssistantPin } from './AssistantPin'
 
 interface SplashParams {
   assistant: dto.UserAssistant
@@ -13,16 +14,19 @@ export const StartChatFromHere = ({ assistant, className }: SplashParams) => {
   return (
     <div className={`flex flex-col ${className}`}>
       <div className="max-h-full overflow-x-hidden flex items-center">
-        <div className="mx-auto flex flex-col gap-3 px-3 pt-12 align-center">
+        <div className="mx-auto flex flex-col gap-3 px-3 pt-12 align-center group">
           <h1 className="text-center">{t('new-chat-title')}</h1>
-          <div className="flex flex-col items-center">
-            <Avatar
-              url={assistant.iconUri ?? undefined}
-              fallback={assistant?.name ?? ''}
-              size="big"
-            ></Avatar>
+          <div className="flex flex-horz self-center">
+            <div className="flex flex-col items-center">
+              <Avatar
+                url={assistant.iconUri ?? undefined}
+                fallback={assistant?.name ?? ''}
+                size="big"
+              ></Avatar>
+              <h3 className="text-center">{assistant?.name}</h3>
+            </div>
+            <AssistantPin assistant={assistant}></AssistantPin>
           </div>
-          <h3 className="text-center">{assistant?.name}</h3>
         </div>
       </div>
